refactor: update camera settings and introduce pose validation
- Changed capture_position_offset type from int to float in camera_settings_data.py
- Added EndEffectorPoint class to define physical points on the robot end-effector
- Introduced validate_pose method in IRobotService and its implementations for pose reachability validation
- Updated robot_service to utilize the new validate_pose method
- Added verify_area_grid method to ICalibrationService for area grid verification
- Enhanced README with details on new features and usage

+++ src/applications/camera_settings/camera_settings_data.py
@@ -11,7 +11,7 @@ class CameraSettingsData:
     width:                    int   = 1280
     height:                   int   = 720
     skip_frames:              int   = 30
-    capture_position_offset:  int   = -4
+    capture_position_offset:  float = -4.0
 
     # Detection
     contour_detection:        bool  = True
